backend/requirements/common.txt
django~=3.0.0

pillow
djangorestframework
djangorestframework-jwt
django-cors-headers
django-taggit
django-taggit-templatetags2